requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
numpy==1.26.0 
pandas==1.3.3
scikit-learn==0.24.2
nltk==3.8.1
streamlit==1.28.0
scikit-learn==1.2.2
python-dotenv==1.0.0
langchain==0.0.326
duckdb==0.9.1
python==3.8.11